requirements.txt

pytesseract==0.3.10
numpy==1.24.3
pandas==2.0.3
Pillow==9.2.0
pip==23.2.1
plotly==5.15.0
streamlit==1.25.0
streamlit_lottie==0.0.5
openpyxl == 3.1.2